速度を実数値に変更

+++ Classes/Character.hpp
@@ -29,8 +29,8 @@ public:
     bool init() override;
 //    void update(float dt) override;
     //    CC_SYNTHESIZE_PASS_BY_REF(cocos2d::Vec2, _acceleration, Acceleration);
-    CC_SYNTHESIZE(int , _oldSpeed, OldSpeed);
-    CC_SYNTHESIZE(int , _speed, Speed);
+    CC_SYNTHESIZE(float , _oldSpeed, OldSpeed);
+    CC_SYNTHESIZE(float , _speed, Speed);
 
     CC_SYNTHESIZE(cocos2d::Vector<cocos2d::SpriteFrame *> , _frames, Frames);
     CREATE_FUNC(Character);
